Artificial Intelligence (AI)/Projects/LangChain/Retrieval Augmented Generation (RAG)/Llama 3 and Groq Inference Engine/requirements.txt
openai
langchain-openai
langchain-core
python-dotenv
langchain_community
pypdf
faiss-cpu
groq
langchain-groq
streamlit